requirements.txt
boto==2.43.0
boto3==1.4.1
django==1.11
django-crispy-forms==1.6.1
django-environ==0.4.0
django-filter==1.0.2
django-storages==1.5.1
djangorestframework==3.6
djangorestframework-filters==0.10.0
django-rest-swagger==2.1.2
faker==0.7.11
freezegun==0.3.8
markdown==2.6.8
model-mommy==1.3.2
pygments==2.2.0
pytest==3.0.7
pytest-django==3.1.2
zappa==0.41.1
